requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0

email-validator==2.1.1
cryptography==41.0.7
